bring over patches to database.rake to prevent production db mishaps

+++ railties/lib/tasks/databases.rake
@@ -93,6 +93,8 @@ namespace :db do
 
   desc 'Drops the database for the current RAILS_ENV'
   task :drop => :load_config do
+    # GITHUB
+    fail "refusing to drop DB in production environment" if RAILS_ENV == 'production'
     config = ActiveRecord::Base.configurations[RAILS_ENV || 'development']
     drop_database(config)
   end
@@ -214,6 +216,8 @@ namespace :db do
   namespace :fixtures do
     desc "Load fixtures into the current environment's database.  Load specific fixtures using FIXTURES=x,y. Load from subdirectory in test/fixtures using FIXTURES_DIR=z. Specify an alternative path (eg. spec/fixtures) using FIXTURES_PATH=spec/fixtures."
     task :load => :environment do
+      # GITHUB
+      fail "refusing to load schema in production environment" if RAILS_ENV == 'production'
       require 'active_record/fixtures'
       ActiveRecord::Base.establish_connection(Rails.env)
       base_dir = ENV['FIXTURES_PATH'] ? File.join(Rails.root, ENV['FIXTURES_PATH']) : File.join(Rails.root, 'test', 'fixtures')
